elastic/requirements.txt
httpx>=0.27.0
mcp>=1.0.0
python-dotenv>=1.0.0
elasticsearch==9.3.0